Organized the notebooks by dividing them into subsections + added more explanatory comments + addressed a few technical issues
The technical issues were related to the notebook 2. In section 2_7_3, walltime, an optional parameters (with default value of 1 hour) was set to 5 hours to prevent failed jobs for large simulations. Another changes was made on the CyberGIS-Compute Service (version 1) side (not in the notebook) to allow for larger data uploads to it (50Mb increased to 1 Gb); before this for large simulations, the CyberGIS-Compute Service  (version 1) could initiat the job due to uploaded file being too large.